refactor media inputs
