+++ Sensus.Shared/UI/Inputs/Input.cs
@@ -564,26 +564,6 @@ namespace Sensus.UI.Inputs
             }
 
             _view = viewContainer;
-
-            if (HasMedia)
-            {
-                MediaView mediaView = new MediaView()
-                {
-                    HorizontalOptions = LayoutOptions.FillAndExpand,
-                    VerticalOptions = LayoutOptions.FillAndExpand
-                };
-
-                MediaView = mediaView;
-
-                StackLayout mediaLayout = new StackLayout
-                {
-                    Children = { mediaView, _view },
-                    Orientation = StackOrientation.Vertical,
-                    HorizontalOptions = LayoutOptions.CenterAndExpand
-                };
-
-                _view = mediaLayout;
-            }
         }
 
         public void Reset()
@@ -648,36 +628,6 @@ namespace Sensus.UI.Inputs
         public override string ToString()
         {
             return _name;
-        }
-
-        [MediaPickerUiProperty("Media:", true, 7)]
-        public MediaObject Media { get; set; }
-
-        [JsonIgnore]
-        public MediaView MediaView { get; private set; }
-
-        public bool HasMedia
-        { 
-            get
-            {
-                return Media != null && string.IsNullOrWhiteSpace(Media.Data) == false;
-            }
-        }
-
-        public async Task InitializeMediaAsync()
-        {
-            if (HasMedia)
-            {
-                await MediaView.SetMediaAsync(Media);
-            }
-        }
-
-        public async Task DisposeMediaAsync()
-        {
-            if (HasMedia)
-            {
-                await MediaView.DisposeMediaAsync();
-            }
         }
     }
 }
